+++ Epoch.net/Constants.cs
@@ -4,6 +4,10 @@ namespace Epoch.net
 {
     public static class Constants
     {
-        public static DateTime UnixEpoch = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
+        public static readonly DateTime UnixEpoch = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
+        public const int MIN_VALUE_INT = -2147483648;
+        public const int MAX_VALUE_INT = 2147483647;
+        public static readonly DateTime MAX_VALUE_DATETIME = new DateTime(2038, 1, 13,3,14,07, DateTimeKind.Utc); // 3:14:07 Tuesday, 19 January 2038 UTC
+        public static readonly DateTime MIN_VALUE_DATETIME = new DateTime(1901, 12, 13, 20, 45, 52, DateTimeKind.Utc); // 20:45:52 Friday, 13 December 1901 UTC
     }
 }