+++ Epoch.net/Exceptions/EpochValueException.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Epoch.net.Exceptions
+{
+    public class EpochValueException : Exception
+    {
+        public EpochValueException(string message)
+            : base(message){}
+    }
+}